modules/BlinkingLedModule/requirements.txt
azure-iothub-device-client~=1.4.3
RPi.GPIO